+++ .github/release-notes/release-notes.md
@@ -2,7 +2,8 @@ Provides common Equality Comparers.
 
 ## ✨ New Features
 
-- [#6](https://github.com/zhofre/test-fabric/issues/6) Added relative and absolute double comparers.
+- [#6](https://github.com/zhofre/test-fabric/issues/6) Added relative and absolute double equality comparers.
+- [#6](https://github.com/zhofre/test-fabric/issues/6) Added object equality comparer.
 
 ## 📦 NuGet